+++ Makefile
@@ -1,4 +1,8 @@
 
+.PHONY: cache
+cache:
+	python3 manage.py createcachetable
+
 .PHONY: clean-pyc
 clean-pyc:
 	find . -name '*.pyc' -exec rm -f {} +
