+++ dashboard/managers/inventory.py
@@ -17,15 +17,11 @@
 #       Languages,
 #       Translation Platforms,
 #       Release Streams, and Branches
-#       Packages
 
 # python
 import io
-import re
-import difflib
-import operator
-from collections import OrderedDict
 from uuid import uuid4
+from collections import OrderedDict
 
 # third party
 import requests
@@ -38,19 +34,15 @@ from django.utils import timezone
 from dashboard.managers.base import BaseManager
 from dashboard.models import (
     TransPlatform, Languages, LanguageSet,
-    ReleaseStream, StreamBranches, Packages, SyncStats
+    ReleaseStream, StreamBranches, SyncStats
 )
 from dashboard.constants import (
-    TRANSPLATFORM_ENGINES, ZANATA_SLUGS, DAMNEDLIES_SLUGS,
-    TRANSIFEX_SLUGS, RELSTREAM_SLUGS, BRANCH_MAPPING_KEYS,
+    ZANATA_SLUGS, DAMNEDLIES_SLUGS, TRANSIFEX_SLUGS, RELSTREAM_SLUGS
 )
-from dashboard.managers.utilities import (
-    parse_project_details_json, parse_ical_file
-)
+from dashboard.managers.utilities import parse_ical_file
 
 
-__all__ = ['InventoryManager', 'SyncStatsManager', 'PackagesManager',
-           'ReleaseBranchManager', 'PackageBranchMapping']
+__all__ = ['InventoryManager', 'SyncStatsManager', 'ReleaseBranchManager']
 
 
 class InventoryManager(BaseManager):
@@ -104,7 +96,7 @@ class InventoryManager(BaseManager):
         """
         locales = self.get_locales()
         if not locales:
-            return
+            return ()
         active_locales = [locale for locale in locales if locale.lang_status]
         inactive_locales = list(set(locales) - set(active_locales))
         aliases = list(filter(lambda locale: locale.locale_alias is not None, locales))
@@ -362,189 +354,13 @@ class SyncStatsManager(BaseManager):
                 locale_translated.append([stats_dict.get('locale', ''), translation_percent])
         return locale_translated
 
-
-class PackagesManager(InventoryManager):
-    """
-    Packages Manager
-    """
-
-    syncstats_manager = SyncStatsManager()
-
-    def get_packages(self, pkgs=None, pkg_params=None):
-        """
-        fetch packages from db
-        """
-        packages = None
-        fields = pkg_params if isinstance(pkg_params, (list, tuple)) else []
-        kwargs = {}
-        if pkgs:
-            kwargs.update(dict(package_name__in=pkgs))
-        try:
-            packages = Packages.objects.only(*fields).filter(**kwargs) \
-                .order_by('-transtats_lastupdated')
-        except Exception as e:
-            self.app_logger(
-                'ERROR', "Packages could not be fetched, details: " + str(e)
-            )
-        return packages
-
-    def is_package_exist(self, package_name):
-        """
-        check package existance
-        """
-        if self.get_packages(pkgs=[package_name]):
-            return True
-        return False
-
-    def update_package(self, package_name, fields):
-        """
-        Update a package with certain fields
-        """
-        try:
-            Packages.objects.filter(package_name=package_name).update(**fields)
-        except Exception as e:
-            self.app_logger(
-                'ERROR', "Package could not be updated, details: " + str(e)
-            )
-
-    def get_relbranch_specific_pkgs(self, release_branch, fields=None):
-        """
-        fetch release branch specific packages from db
-        """
-        packages = ()
-        fields_required = fields if fields else ()
-        try:
-            packages = Packages.objects.only(*fields_required) \
-                .filter(release_branch_mapping__has_key=release_branch) \
-                .order_by('-transtats_lastupdated')
-        except Exception as e:
-            self.app_logger(
-                'ERROR', ("release branch specific package could not be fetched for " +
-                          release_branch + " release branch, details: " + str(e))
-            )
-        return packages
-
-    def count_packages(self):
-        """
-        packages count
-        """
-        packages = self.get_packages()
-        return packages.count() if packages else 0
-
-    def get_package_name_tuple(self, t_status=False, check_mapping=False):
-        """
-        returns (package_name, upstream_name) tuple (only sync'd ones)
-        """
-        packages = self.get_packages()
-        name_list = [(package.package_name, package.upstream_name) for package in packages]
-        if t_status:
-            name_list = [(package.package_name, package.upstream_name) for package in packages
-                         if package.transtats_lastupdated or package.upstream_lastupdated]
-        elif check_mapping:
-            name_list = [(package.package_name, package.package_name) for package in packages
-                         if package.transtats_lastupdated and package.package_name_mapping]
-        return tuple(sorted(name_list))
-
-    def _get_project_details(self, transplatform, package_name):
-        """
-        Get platform-wise project details
-        """
-        resp_dict = None
-        transplatform_url = None
-        if transplatform.engine_name == TRANSPLATFORM_ENGINES[0]:
-            transplatform_url = transplatform.api_url + "/module/" + package_name + "/"
-            resp_dict = self.api_resources.fetch_project_details(
-                transplatform.engine_name, transplatform.api_url, package_name
-            )
-        elif transplatform.engine_name == TRANSPLATFORM_ENGINES[1]:
-            resp_dict = self.api_resources.fetch_project_details(
-                transplatform.engine_name, transplatform.api_url, package_name,
-                **dict(ext=True, auth_user=transplatform.auth_login_id, auth_token=transplatform.auth_token_key)
-            )
-            if resp_dict:
-                tx_org_slug = resp_dict['organization']['slug']
-                transplatform_url = transplatform.api_url + "/" + tx_org_slug + "/" + package_name
-        elif transplatform.engine_name == TRANSPLATFORM_ENGINES[2]:
-            transplatform_url = transplatform.api_url + "/project/view/" + package_name
-            resp_dict = self.api_resources.fetch_project_details(
-                transplatform.engine_name, transplatform.api_url, package_name,
-                **dict(auth_user=transplatform.auth_login_id, auth_token=transplatform.auth_token_key)
-            )
-        return transplatform_url, resp_dict
-
-    def add_package(self, **kwargs):
-        """
-        add package to db
-        :param kwargs: dict
-        :return: boolean
-        """
-        required_params = ('package_name', 'upstream_url', 'transplatform_slug', 'release_streams')
-        if not set(required_params) <= set(kwargs.keys()):
-            return
-
-        if not (kwargs['package_name'] and kwargs['upstream_url']):
-            return
-
-        try:
-            # derive translation platform project URL
-            platform = TransPlatform.objects.only('engine_name', 'api_url') \
-                .filter(platform_slug=kwargs['transplatform_slug']).get()
-            kwargs['transplatform_url'], resp_dict = \
-                self._get_project_details(platform, kwargs['package_name'])
-            if resp_dict:
-                # save project details in db
-                kwargs['package_details_json'] = resp_dict
-                kwargs['details_json_lastupdated'] = timezone.now()
-                if kwargs.get('update_stats') == 'stats':
-                    kwargs.pop('update_stats')
-                    # fetch project_version stats from translation platform and save in db
-                    project, versions = parse_project_details_json(platform.engine_name, resp_dict)
-                    for version in versions:
-                        resp_dict = {}
-                        if platform.engine_name == TRANSPLATFORM_ENGINES[0]:
-                            # this is a quick fix for chinese in DamnedLies modules
-                            locales = [locale.locale_alias if 'zh' not in locale.locale_id else locale.locale_id
-                                       for locale in self.get_locales(only_active=True)]
-                            locales_stats_list = []
-                            for locale in locales:
-                                locale_stats = self.api_resources.fetch_translation_statistics(
-                                    platform.engine_name, platform.api_url, locale, version,
-                                    **dict(package_name=kwargs['package_name'] or project)
-                                )
-                                if locale_stats:
-                                    locales_stats_list.append(locale_stats)
-                            resp_dict['id'] = version
-                            resp_dict['stats'] = locales_stats_list
-                        else:
-                            resp_dict = self.api_resources.fetch_translation_statistics(
-                                platform.engine_name, platform.api_url, project, version,
-                                **dict(auth_user=platform.auth_login_id, auth_token=platform.auth_token_key)
-                            )
-                        if resp_dict:
-                            if self.save_version_stats(project, version, resp_dict, platform.engine_name):
-                                kwargs['transtats_lastupdated'] = timezone.now()
-            if 'update_stats' in kwargs:
-                del kwargs['update_stats']
-
-            kwargs['transplatform_slug'] = platform
-            kwargs['transplatform_name'] = kwargs['package_name']
-            kwargs['upstream_name'] = kwargs['upstream_url'].split('/')[-1]
-            # save in db
-            new_package = Packages(**kwargs)
-            new_package.save()
-        except:
-            # log event, pass for now
-            # todo - implement error msg handling
-            return False
-        else:
-            return True
-
     def save_version_stats(self, project, version, stats_json, stats_source):
         """
         Save version's translation stats in db
         :param project: transplatform project
         :param version: transplatform project's version
         :param stats_json: translation stats dict
+        :param stats_source: platform engine or build system
         :return: boolean
         """
 
@@ -576,298 +392,6 @@ class PackagesManager(InventoryManager):
         else:
             return True
         return False
-
-    def _get_project_ids_names(self, engine, projects):
-        ids = []
-        names = []
-        if isinstance(projects, list) and engine == TRANSPLATFORM_ENGINES[0]:
-            for project in projects:
-                ids.append(project['fields']['name'])
-        elif isinstance(projects, dict) and engine == TRANSPLATFORM_ENGINES[1]:
-            ids.append(projects.get('slug'))
-            names.append(projects.get('name'))
-        elif isinstance(projects, list) and engine == TRANSPLATFORM_ENGINES[2]:
-            for project in projects:
-                ids.append(project['id'])
-                names.append(project['name'])
-        return ids, names
-
-    def validate_package(self, **kwargs):
-        """
-        Validates existence of a package at a transplatform
-        :param kwargs: dict
-        :return: package_name: str, Boolean
-        """
-        if not (kwargs.get('package_name')):
-            return
-        package_name = kwargs['package_name']
-        transplatform_fields = ('engine_name', 'api_url', 'projects_json',
-                                'auth_login_id', 'auth_token_key')
-        # get transplatform projects from db
-        platform = TransPlatform.objects.only(*transplatform_fields) \
-            .filter(platform_slug=kwargs['transplatform_slug']).get()
-        projects_json = platform.projects_json
-        # if not found in db, fetch transplatform projects from API
-        if not projects_json:
-            response_dict = None
-            auth_dict = dict(
-                auth_user=platform.auth_login_id, auth_token=platform.auth_token_key
-            )
-            if platform.engine_name == TRANSPLATFORM_ENGINES[1]:
-                response_dict = self.api_resources.fetch_project_details(
-                    platform.engine_name, platform.api_url, package_name.lower(), **auth_dict
-                )
-            elif platform.engine_name == TRANSPLATFORM_ENGINES[0] or \
-                    platform.engine_name == TRANSPLATFORM_ENGINES[2]:
-                response_dict = self.api_resources.fetch_all_projects(
-                    platform.engine_name, platform.api_url, **auth_dict
-                )
-                # save all_projects_json in db - faster validation next times
-                # except transifex, as there we have project level details
-                TransPlatform.objects.filter(api_url=platform.api_url).update(
-                    projects_json=response_dict, projects_lastupdated=timezone.now()
-                )
-            if response_dict:
-                projects_json = response_dict
-        ids, names = self._get_project_ids_names(platform.engine_name, projects_json)
-        if package_name in ids:
-            return package_name
-        elif package_name in names:
-            return ids[names.index(package_name)]
-        else:
-            return False
-
-    def _get_lang_id_name_dict(self, release_branch=None):
-        """
-        Generates {(locale, alias): language_name} dict
-        """
-        active_locales = self.get_locales(
-            pick_locales=self.get_relbranch_locales(release_branch)
-        ) if release_branch else self.get_locales_set()[0]
-        lang_id_name = dict([((lang.locale_id, lang.locale_alias), lang.lang_name)
-                             for lang in active_locales])
-        return OrderedDict(sorted(lang_id_name.items(), key=operator.itemgetter(1)))
-
-    def get_trans_stats(self, package_name, apply_branch_mapping=False, specify_branch=None):
-        """
-        fetch stats of a package for all enabled languages
-        :param package_name: str
-        :param apply_branch_mapping: boolean
-        :return: dict {project_version: stats_dict}
-        """
-        trans_stats_dict = OrderedDict()
-        package_desc = ''
-        # 1st, get active locales for which stats are to be shown
-        #      or, choose release branch specific locales
-        lang_id_name = self._get_lang_id_name_dict(specify_branch) \
-            if apply_branch_mapping and specify_branch else self._get_lang_id_name_dict()
-        # 2nd, filter stats json for required locales
-        if self.is_package_exist(package_name):
-            package_details = self.get_packages([package_name]).get()
-            if not (package_details.transtats_lastupdated or package_details.upstream_lastupdated):
-                return lang_id_name, trans_stats_dict, package_desc
-            if package_details.package_details_json and package_details.package_details_json.get('description'):
-                package_desc = package_details.package_details_json['description']
-            pkg_stats_versions = self.syncstats_manager.get_sync_stats(pkgs=[package_name])
-            for pkg_stats_version in pkg_stats_versions:
-                trans_stats_list, missing_locales = \
-                    self.syncstats_manager.filter_stats_for_required_locales(
-                        package_details.transplatform_slug_id,
-                        pkg_stats_version.stats_raw_json, list(lang_id_name)
-                    )
-                if 'test' not in pkg_stats_version.project_version and 'extras' not in pkg_stats_version.project_version:
-                    trans_stats_dict[pkg_stats_version.project_version] = \
-                        self.syncstats_manager.extract_locale_translated(package_details.transplatform_slug_id,
-                                                                         trans_stats_list)
-            if apply_branch_mapping and package_details.release_branch_mapping:
-                branch_mapping = package_details.release_branch_mapping
-                for relbranch, branch_mapping in branch_mapping.items():
-                    trans_stats_dict[relbranch] = trans_stats_dict.get(branch_mapping.get(BRANCH_MAPPING_KEYS[0]), [])
-        return lang_id_name, trans_stats_dict, package_desc
-
-    def get_upstream_stats(self, package):
-        """
-        fetch upstream stats of a package for all enabled locales
-        :param package: str
-        :return:upstream_trans_stats: list
-        """
-        upstream_trans_stats = []
-        if not package:
-            return upstream_trans_stats
-        lang_id_name = self._get_lang_id_name_dict()
-        package_details = self.get_packages([package], ['upstream_latest_stats']).get()
-        upstream_stats = package_details.upstream_latest_stats if package_details else {}
-        if upstream_stats:
-            selected_locales = []
-            for locale in list(upstream_stats.keys()):
-                selected_locales.extend([locale for locale_tuple in list(lang_id_name.keys())
-                                         if locale in locale_tuple])
-            for s_locale in selected_locales:
-                upstream_trans_stats.append(
-                    [s_locale, upstream_stats.get(s_locale, {}).get('translated_percent')]
-                )
-        return upstream_trans_stats
-
-    def _get_pkg_and_ext(self, package_name):
-        package = self.get_packages([package_name]).get()
-        # extension for Transifex should be true, otherwise false
-        extension = (True if package.transplatform_slug.engine_name == TRANSPLATFORM_ENGINES[0] else False)
-        return package, extension
-
-    def sync_update_package_details(self, package_name):
-        """
-        Sync with translation platform and update details in db for a package
-        :param package_name: str
-        :return: boolean
-        """
-        update_pkg_status = False
-        package, ext = self._get_pkg_and_ext(package_name)
-        platform = package.transplatform_slug
-        project_details_response_dict = self.api_resources.fetch_project_details(
-            platform.engine_name, platform.api_url, package_name,
-            **(dict(auth_user=platform.auth_login_id, auth_token=platform.auth_token_key))
-        )
-        if project_details_response_dict:
-            try:
-                Packages.objects.filter(transplatform_url=package.transplatform_url).update(
-                    package_details_json=project_details_response_dict,
-                    details_json_lastupdated=timezone.now()
-                )
-            except Exception as e:
-                self.app_logger(
-                    'ERROR', "Package update failed, details: " + str(e))
-            else:
-                update_pkg_status = True
-        return update_pkg_status
-
-    def sync_update_package_stats(self, package_name):
-        """
-        Sync with translation platform and update trans stats in db for a package
-        :param package_name: str
-        :return: boolean
-        """
-        update_stats_status = False
-        package, ext = self._get_pkg_and_ext(package_name)
-        project, versions = parse_project_details_json(
-            package.transplatform_slug.engine_name, package.package_details_json
-        )
-        for version in versions:
-            proj_trans_stats_response_dict = {}
-            if package.transplatform_slug.engine_name == TRANSPLATFORM_ENGINES[0]:
-                # this is a quick fix for chinese in DamnedLies modules
-                locales = [locale.locale_alias if 'zh' not in locale.locale_id else locale.locale_id
-                           for locale in self.get_locales(only_active=True)]
-                locales_stats_list = []
-                for locale in locales:
-                    locale_stats = self.api_resources.fetch_translation_statistics(
-                        package.transplatform_slug.engine_name, package.transplatform_slug.api_url,
-                        locale, version, **dict(package_name=package_name)
-                    )
-                    if locale_stats:
-                        locales_stats_list.append(locale_stats)
-                proj_trans_stats_response_dict.update({"id": version, "stats": locales_stats_list})
-            else:
-                proj_trans_stats_response_dict = self.api_resources.fetch_translation_statistics(
-                    package.transplatform_slug.engine_name, package.transplatform_slug.api_url, project, version,
-                    **dict(auth_user=package.transplatform_slug.auth_login_id,
-                           auth_token=package.transplatform_slug.auth_token_key)
-                )
-            if proj_trans_stats_response_dict:
-                if self.save_version_stats(
-                        project, version, proj_trans_stats_response_dict, package.transplatform_slug.engine_name
-                ):
-                    Packages.objects.filter(transplatform_url=package.transplatform_url).update(
-                        transtats_lastupdated=timezone.now())
-                    update_stats_status = True
-        # this makes sense if we create branch-mapping just after package sync
-        self.build_branch_mapping(package_name)
-        return update_stats_status
-
-    @staticmethod
-    def get_pkg_branch_mapping(pkg):
-        return PackageBranchMapping(pkg).branch_mapping
-
-    def build_branch_mapping(self, package_name):
-        """
-        Creates Branch mapping for the Package
-        :param package_name: str
-        :return: boolean
-        """
-        kwargs = {}
-        branch_mapping_dict = self.get_pkg_branch_mapping(package_name)
-        kwargs['package_name_mapping'] = {package_name: ''}
-        kwargs['release_branch_mapping'] = branch_mapping_dict
-        kwargs['mapping_lastupdated'] = timezone.now()
-        try:
-            Packages.objects.filter(package_name=package_name).update(**kwargs)
-        except Exception as e:
-            self.app_logger(
-                'ERROR', "Package branch mapping could not be saved, details: " + str(e))
-        else:
-            return True
-        return False
-
-    def refresh_package(self, package_name):
-        """
-        Sync Stats and Rebuild Mapping
-        :param package_name: str
-        :return: boolean
-        """
-        if not package_name:
-            return False
-
-        status = []
-        steps = (
-            self.sync_update_package_details,
-            self.sync_update_package_stats
-        )
-
-        for method in steps:
-            status.append(method(package_name))
-        return True if [i for i in status if i] else False
-
-    def _formats_stats_diff(self, lang_sequence, stats_diff_dict):
-        lang_dict = {}
-        [lang_dict.update({index: lang}) for index, lang in lang_sequence]
-        formatted_stats_dict = OrderedDict()
-        for branch, stats_diff in stats_diff_dict.items():
-            formatted_stats_dict[branch] = {}
-            diff = list(stats_diff)
-            stats_group = OrderedDict()
-            for x in diff:
-                stats_group.setdefault(x[0], []).append(x[1])
-            for key, value in stats_group.items():
-                formatted_stats_dict[branch][lang_dict.get(key)] = \
-                    value[1] - value[0] if isinstance(value, list) and len(value) > 1 else 0
-        return formatted_stats_dict
-
-    def calculate_stats_diff(self, package, graph_ready_stats, pkg_branch_map):
-        """
-        Calculates and stores translation stats differences
-        :param package: str
-        :param graph_ready_stats: dict
-        :param pkg_branch_map: dict
-        :return: languages list and stats diff dict
-        """
-        stats_diff_dict = OrderedDict()
-        stats_data = graph_ready_stats.get('graph_data', {})
-        languages = graph_ready_stats.get('ticks', [])
-        for branch, mapping in pkg_branch_map.items():
-            platform_stats = stats_data.get(mapping[BRANCH_MAPPING_KEYS[0]], [])
-            buildsys_stats = stats_data.get(
-                mapping[BRANCH_MAPPING_KEYS[1]] + " - " + mapping[BRANCH_MAPPING_KEYS[2]], []
-            )
-            # Ignoring after-decimal differences for now
-            first_set = set(map(tuple, [(index, round(stat)) for index, stat in platform_stats]))
-            second_set = set(map(tuple, [(index, round(stat)) for index, stat in buildsys_stats]))
-            if first_set and second_set:
-                stats_diff_dict[branch] = first_set.symmetric_difference(second_set)
-            else:
-                raise Exception("Make sure all mapped versions/tags are sync'd for stats.")
-        polished_stats_diff = self._formats_stats_diff(languages, stats_diff_dict)
-        if package and polished_stats_diff:
-            self.update_package(package, {'stats_diff': polished_stats_diff})
-        return polished_stats_diff
 
 
 class ReleaseBranchManager(InventoryManager):
@@ -1047,209 +571,3 @@ class ReleaseBranchManager(InventoryManager):
             return False
         else:
             return True
-
-
-class PackageBranchMapping(object):
-    """
-    Creates Branch Mapping
-    """
-    package = None
-    package_name = None
-    original_versions = None
-    transplatform_versions = None
-    release_branches_dict = None
-    release_branches_list = None
-    release_streams_list = None
-    release_build_tags_dict = None
-
-    syncstats_manager = SyncStatsManager()
-    relbranch_manager = ReleaseBranchManager()
-
-    def __init__(self, package_name):
-        self.package_name = package_name
-        try:
-            self.package = Packages.objects.filter(package_name=self.package_name).first()
-        except Exception as e:
-            # log event, passing for now
-            pass
-        else:
-            pkg_stats = self.syncstats_manager.get_sync_stats(pkgs=[package_name], fields=('project_version',))
-            self.original_versions = [pkg_stat.project_version for pkg_stat in pkg_stats]
-            self.transplatform_versions = [
-                pkg_stat.project_version.lower() for pkg_stat in self.syncstats_manager.get_sync_stats(
-                    pkgs=[package_name], fields=('project_version',), sources=TRANSPLATFORM_ENGINES
-                ) if isinstance(pkg_stat.project_version, str)]
-            self.release_branches_dict = self.relbranch_manager.get_branches_of_relstreams(self.package.release_streams)
-            self.release_branches_list = []
-            self.release_streams_list = []
-            self.release_build_tags_dict = {}
-            [self.release_branches_list.extend(branches) for _, branches in self.release_branches_dict.items()]
-            [self.release_streams_list.extend([stream]) for stream, _ in self.release_branches_dict.items()
-             if stream not in self.release_streams_list]
-            for release_stream in self.release_streams_list:
-                build_tags = self.relbranch_manager.get_relstream_build_tags(stream_slug=release_stream)
-                if build_tags:
-                    self.release_build_tags_dict.update(build_tags)
-
-    def _get_stream_branch_belongs_to(self, branch):
-        """
-        Get stream to which a branch belongs to
-        """
-        for stream, branches in self.release_branches_dict.items():
-            if branch in branches:
-                return stream
-        return ''
-
-    def _sort_and_match_version_nm(self, release_branch, from_branches):
-        """
-        Sort versions matching stream and then try to match version number
-        """
-        belonging_stream = self._get_stream_branch_belongs_to(release_branch).lower()
-        relevant_versions = [version for version in from_branches if belonging_stream in version]
-        version_numbers = re.findall(r'\d+', release_branch)
-        if len(version_numbers) >= 1:
-            expected_version = [relevant_version for relevant_version in relevant_versions
-                                if version_numbers[0] in relevant_version]
-            if len(expected_version) >= 1:
-                return True, expected_version[0]
-        return False, ''
-
-    def _check_release_version(self, branch, versions, shortform=None):
-        """
-        Match version numbers
-        """
-        short_form = shortform if shortform else ''
-        version_numbers = re.findall(r'\d+', branch)
-        if len(version_numbers) >= 1:
-            short_form += str(version_numbers[0])
-        return [version for version in versions if short_form in version]
-
-    def _compare_with_short_names(self, release_branch, from_branches):
-        """
-        Try short forms combined with version number
-        """
-        belonging_stream = self._get_stream_branch_belongs_to(release_branch).lower()
-        short_form = belonging_stream
-        if belonging_stream == RELSTREAM_SLUGS[1]:
-            short_form = 'f'
-        expected_version = self._check_release_version(
-            release_branch, from_branches, shortform=short_form
-        )
-        if len(expected_version) >= 1:
-            return True, expected_version[0]
-        return False, ''
-
-    def _return_original_version(self, matched_version):
-        required_version = [version for version in self.original_versions
-                            if version.lower() == matched_version]
-        return required_version[0] if len(required_version) == 1 else matched_version
-
-    def calculate_branch_mapping(self, branch, from_branches):
-        """
-        Calculates branch mapping for a branch
-
-        Algorithm:
-        1. try with difflib, check version and return first found
-        2. sort versions matching stream and then try to match version number
-        3. try short forms combined with version number
-        4. todo: fetch release dates and try matching with that
-        5. try finding default version: 'master' may be
-        6. if nothing works, return blank
-        """
-        match1 = difflib.get_close_matches(branch, from_branches)
-        if len(match1) >= 1:
-            match_found = self._check_release_version(branch, match1)
-            if len(match_found) >= 1:
-                return self._return_original_version(match_found[0])
-
-        status1, match2 = self._sort_and_match_version_nm(branch, from_branches)
-        if status1:
-            return self._return_original_version(match2)
-
-        status2, match3 = self._compare_with_short_names(branch, from_branches)
-        if status2:
-            return self._return_original_version(match3)
-
-        probable_branches = ['default', 'master', 'head', 'rawhide',
-                             'devel', 'core', self.package_name]
-
-        for branch in probable_branches:
-            if branch in from_branches:
-                return self._return_original_version(branch)
-        return ''
-
-    def branch_stats(self, release_branch):
-        """
-        Generates stats for a specific release branch
-        :param release_branch: release branch slug
-        :return: dict
-        """
-        if not release_branch:
-            return {}
-        transplatform_version = self.branch_mapping.get(
-            release_branch, {}).get(BRANCH_MAPPING_KEYS[0])
-        pkg_stats_query_set = self.syncstats_manager.get_sync_stats(
-            pkgs=[self.package_name], versions=[transplatform_version]
-        )
-        pkg_stats_json = pkg_stats_query_set.first().stats_raw_json \
-            if pkg_stats_query_set else {}
-        return pkg_stats_json
-
-    @property
-    def versions(self):
-        return self.transplatform_versions
-
-    @property
-    def release_branches(self):
-        return self.release_branches_list
-
-    @property
-    def branch_mapping(self):
-        """
-        Creates branch mapping based on rule
-        :return: dict
-        """
-        branch_mapping_dict = {}
-        required_params = (self.transplatform_versions, self.release_branches_dict,
-                           self.release_streams_list, self.release_build_tags_dict)
-        valid_params = [param for param in required_params if param]
-
-        if len(required_params) == len(valid_params):
-            for stream, branches in self.release_branches_dict.items():
-                for branch in branches:
-                    branch_mapping_dict[branch] = {}
-                    branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[0]] = \
-                        self.calculate_branch_mapping(branch, self.transplatform_versions)
-                    branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[1]] = \
-                        self.relbranch_manager.get_relstream_buildsys(stream)
-                    branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[2]] = \
-                        self.calculate_branch_mapping(branch, sorted(
-                            self.release_build_tags_dict[stream]))
-
-                    if not branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[0]] \
-                            and self.package.transplatform_slug_id in DAMNEDLIES_SLUGS:
-                        release_stream = self.relbranch_manager.get_release_streams(
-                            built=branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[1]],
-                            fields=('relstream_server',))
-                        if release_stream:
-                            release_stream_hub_url = release_stream.get().relstream_server or ''
-                            build_info = self.relbranch_manager.api_resources.build_info(
-                                hub_url=release_stream_hub_url,
-                                tag=branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[2]],
-                                pkg=self.package_name
-                            )
-                            if build_info and isinstance(build_info, list) and len(build_info) > 0:
-                                version_from_latest_build = build_info[0].get('version')
-                                seek_version = "-".join(version_from_latest_build.split('.')[0:2])
-                                for version in self.transplatform_versions:
-                                    if seek_version in version:
-                                        branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[0]] = version
-                                # seek next (rounded) version
-                                if not branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[0]]:
-                                    version_x, version_y = version_from_latest_build.split('.')[0:2]
-                                    version_y = str(int(5 * round(float(version_y) / 5)))
-                                    seek_version = "-".join((version_x, version_y))
-                                    for version in self.transplatform_versions:
-                                        if seek_version in version:
-                                            branch_mapping_dict[branch][BRANCH_MAPPING_KEYS[0]] = version
-        return branch_mapping_dict
