+++ dashboard/managers/upstream.py
@@ -26,7 +26,7 @@ from django.utils import timezone
 from dashboard.constants import TS_JOB_TYPES
 from dashboard.managers.base import BaseManager
 from dashboard.managers.jobs import JobManager
-from dashboard.managers.inventory import PackagesManager
+from dashboard.managers.packages import PackagesManager
 
 
 __all__ = ['UpstreamManager']
@@ -141,8 +141,14 @@ class UpstreamManager(BaseManager):
         )
 
         for po_file in trans_files:
-            po = polib.pofile(po_file)
-            if po:
+            try:
+                po = polib.pofile(po_file)
+            except Exception as e:
+                self.job_manager.log_json[SUBJECT].update(
+                    {str(datetime.now()): "Exception: " + str(e)}
+                )
+                self.job_manager.job_result = False
+            else:
                 temp_trans_stats = {}
                 locale = po_file.split('/')[-1].split('.')[0]
                 temp_trans_stats[locale] = {}
