+++ dashboard/templates/stats/_stats_diff.html
@@ -3,23 +3,35 @@
     <div class="panel panel-default">
       <div class="panel-heading" style="padding-bottom:20px">
           <span style="font-size:large">Statistics Differences&nbsp;&nbsp;
-              <small class="text-muted">(<i>Translation Platform Vs Build System</i>)</small></span>
+              <small class="text-muted">(<i>Translation Platform Vs Build System</i>)</small>
+          </span>
       </div>
-      <table class="table table-hover table-bordered" cellspacing="0">
+      <table class="table table-bordered" cellspacing="0">
         <thead>
             <tr>
-                <th>Release Stream Branch</th>
+                <th width="12%">Release Branch</th>
                 <th>Languages <small>(which are out of sync: <span class="text-danger">needs attention</span>)</small></th>
             </tr>
         </thead>
         <tbody>
-            {% for branch, langs in stats_diff.items %}
+            {% for branch, lang_diff in stats_diff.items %}
             <tr>
                 <td>{{ branch }}</td>
-                <td><span class="text-info">{{ langs|join_by:", " }}</span></td>
+                <td>
+                    {% for lang, diff in lang_diff.items %}
+                        {% if diff == 0 %}
+                            <span class="text-info">{{ lang }}</span>&nbsp;&nbsp;
+                        {% elif diff > 0 %}
+                            <span class="text-danger" data-toggle="tooltip" title="{{ diff }}%">{{ lang }}</span>&nbsp;&nbsp;
+                        {% endif %}
+                    {% endfor %}
+                </td>
             </tr>
             {% endfor %}
         </tbody>
       </table>
+      <div class="panel-footer text-muted">
+          Languages in red color differ in stats with Translation Platform. <i>So, either push or pull is not latest.</i> Other languages could <b>not</b> make 100% in the build system.
+      </div>
     </div>
 {% endif %}
