+++ devel/ansible/roles/setup/tasks/main.yml
@@ -11,7 +11,7 @@
   shell: cp /workspace/transtats/settings/keys.json.example /workspace/transtats/settings/keys.json
 
 - name: Setup DB Schema
-  shell: cd /workspace; make migrate
+  shell: cd /workspace; make migrate; make cache
 
 - name: Load Initial Data
   shell: runuser -l postgres -c "psql -d {{ db_name }} -a -f /workspace/deploy/docker/data/initial.sql"
